calisabor/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
Django==3.2.12
django-allauth==0.49.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
djongo==1.3.6
idna==3.3
oauthlib==3.2.0
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.3
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.9
